Fix ValueError: Buffer dtype mismatch on 64-bit Linux systems
Change np.int32 to np.int64 in _to_c_long_array and node/edge index
arrays to match Cython cyaccess extension expectations on 64-bit Linux.

+++ pandana/network.py
@@ -153,8 +153,8 @@ class Network:
 
     @staticmethod
     def _to_c_long_array(values):
-        # cyaccess expects C long buffers; on Windows that maps to 32-bit ints.
-        return np.asarray(values, dtype=np.int32)
+        # cyaccess expects C long buffers; on 64-bit Linux that maps to 64-bit ints.
+        return np.asarray(values, dtype=np.int64)
 
     @property
     def aggregations(self):
